+++ packages/nova/src/rules/eslint/index.ts
@@ -1,0 +1,1 @@
+export { default as noLoggerDev } from './no-logger-dev.js';