+++ packages/nova/src/types/tests/lib/utility.test.d.ts
@@ -1,0 +1,8 @@
+/**
+ * Discover paths with file.
+ *
+ * @since 1.0.0
+ */
+export type DiscoverPathsWithFileOriginalCwd = string;
+
+export type DiscoverPathsWithFileSandboxRoot = string;